+++ lib/core/domain/pentest_report_controller/pentest_report_controller_cubit.dart
@@ -1,7 +1,6 @@
 import 'package:bloc/bloc.dart';
 import 'package:meta/meta.dart';
-import 'package:net_runner/core/data/logger.dart';
-import 'package:net_runner/core/data/task_report_serial/pentest_report_serial.dart';
+import 'package:net_runner/core/domain/api/models/task_report_serial/pentest/pentest_report_serial.dart';
 
 part 'pentest_report_controller_state.dart';
