+++ lib/features/scanning/presentation/scanning_pg.dart
@@ -4,8 +4,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:graphic/graphic.dart';
 import 'package:net_runner/core/data/logger.dart';
-import 'package:net_runner/core/data/task_report_serial/pentest_report_serial.dart';
 import 'package:net_runner/core/domain/api/api_bloc.dart';
+import 'package:net_runner/core/domain/api/models/task_report_serial/pentest/pentest_report_serial.dart';
 import 'package:net_runner/core/domain/pentest_report_controller/pentest_report_controller_cubit.dart';
 import 'package:net_runner/core/domain/task_list/task_list_cubit.dart';
 import 'package:net_runner/features/scanning/presentation/create_scan_page.dart';
